More robust thread manager role handling

+++ modules_tavare/Module.Infrastructure.ThreadManager.js
@@ -6,7 +6,15 @@ Module
   .addEvent("threadCreate", async (ThreadChannel, newlyCreated) => {
     if (ThreadChannel.sendable && newlyCreated && !ThreadChannel.locked && ThreadChannel.type == "GUILD_PUBLIC_THREAD") {
       let message = await ThreadChannel.send({ content: `Adding staff to the thread, one moment please.` })
-      await message.edit({ content: `Adding <@&${Module.config.snowflakes.roles.Admin}><@&${Module.config.snowflakes.roles.Moderator}><@&${Module.config.snowflakes.roles.CommunityGuide}> to the thread`, allowedMentions: { parse: ['roles'] } });
+
+      let role_string = "";
+      if (Module.config.snowflakes.roles.Moderator) role_string += `<@&${Module.config.snowflakes.roles.Moderator}>`
+      if (Module.config.snowflakes.roles.Admin) role_string += `<@&${Module.config.snowflakes.roles.Admin}>`
+      if (Module.config.snowflakes.roles.CommunityGuide) role_string += `<@&${Module.config.snowflakes.roles.CommunityGuide}>`
+      if (role_string == "") return;
+
+
+      await message.edit({ content: `Adding ${role_string} to the thread`, allowedMentions: { parse: ['roles'] } });
       await message.delete()
     }
   });
